Complete init setup

+++ database/migrations/2020_10_29_160644_create_student_maps_table.php
@@ -15,7 +15,7 @@ class CreateStudentMapsTable extends Migration
     {
         Schema::create('student_maps', function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->unsignedInteger('student_id');
+            $table->unsignedBigInteger('student_id');
             $table->string('email')->unique();
             $table->string('username')->nullable();
             $table->timestamps();
